Add Mock Chart Data abstraction + minor changes for abstraction
