+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockChartLayout.java
@@ -34,6 +34,7 @@ public class MockChartLayout extends MockLayout {
     @Override
     boolean onDrop(MockComponent source, int x, int y, int offsetX, int offsetY) {
         if (source instanceof MockLineChartData) {
+            MockLineChartData chartData = (MockLineChartData) source;
             container.addComponent(source);
             return true;
         }
