+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockContainer.java
@@ -256,7 +256,8 @@ public abstract class MockContainer extends MockVisibleComponent implements Drop
     if (component instanceof MockVisibleComponent) {
       // Sprites are only allowed on Canvas, not other containers.
       // Map features are only allowed on Map, not other containers.
-      if (!(component instanceof MockSprite) && !(component instanceof MockMapFeature) && !(component instanceof MockLineChartData)) {
+      // Chart Data components are only allowed on Charts, not other containers.
+      if (!(component instanceof MockSprite) && !(component instanceof MockMapFeature) && !(component instanceof MockChartData)) {
         return true;
       }
     }
