+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockLineChartData.java
@@ -3,11 +3,8 @@ package com.google.appinventor.client.editor.simple.components;
 import com.google.appinventor.client.editor.simple.SimpleEditor;
 import com.google.gwt.user.client.ui.InlineHTML;
 
-public class MockLineChartData extends MockVisibleComponent {
+public class MockLineChartData extends MockChartData {
     public static final String TYPE = "LineChartData";
-
-    // Temporary placeholder for the Chart Data image
-    private InlineHTML labelWidget;
 
     /**
      * Creates a new MockLineChartData component.
@@ -16,10 +13,5 @@ public class MockLineChartData extends MockVisibleComponent {
      */
     public MockLineChartData(SimpleEditor editor) {
         super(editor, TYPE, images.label());
-
-        labelWidget = new InlineHTML();
-        labelWidget.setStylePrimaryName("ode-SimpleMockComponent");
-        labelWidget.setText("LINE CHART DATA");
-        initComponent(labelWidget);
     }
 }
